requirements.txt
streamlit>=1.32.0
pytest>=7.0

# AI features (new)
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0